requirements.txt
# Strands Agents
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Bedrock AgentCore (Memory統合を含む)
bedrock-agentcore[strands-agents]>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0

# LINE Bot SDK
line-bot-sdk>=3.0.0

# Web Framework
flask>=3.0.0

# MCP
mcp>=1.0.0

# AWS
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Tavily Search（検索機能）
tavily-python>=0.3.0
